Create pathtools.py
